pyproject.toml
[project]
name = "smart-post-purchase-ai-guardian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain-core>=0.3.78",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.9",
    "numpy>=2.3.3",
    "openai>=2.2.0",
    "pillow>=11.3.0",
    "pinecone>=7.3.0",
    "pinecone-client>=6.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "supabase>=2.21.1",
]
